Add functionality for adding a user to the Database

+++ AutoShop.Presentation/Controllers/UserController.cs
@@ -1,4 +1,4 @@
-﻿using AutoShop.Domain.Enum;
+﻿using AutoShop.Domain.Extensions;
 using AutoShop.Domain.ViewModels.User;
 using AutoShop.Service.Interfaces;
 using Microsoft.AspNetCore.Authorization;
@@ -17,14 +17,7 @@ namespace AutoShop.Presentation.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetUser(int id)
-        {
-            var response = await _userService.GetUserAsync(id);
-            return PartialView("GetUser", response.Data);
-        }
-
-        [HttpGet]
-        public async Task<IActionResult> Details()
+        public async Task<IActionResult> GetUsers()
         {
             var response = await _userService.GetUsersAsync();
             if (response.StatusCode == Domain.Enum.StatusCode.Ok)
@@ -33,21 +26,60 @@ namespace AutoShop.Presentation.Controllers
             return RedirectToAction("Index", "Home");
         }
 
+        [HttpGet]
+        public async Task<IActionResult> GetUser(int id)
+        {
+            if (ModelState.IsValid)
+            {
+                var response = await _userService.GetUserAsync(id);
+                if (response.StatusCode == Domain.Enum.StatusCode.Ok)
+                    return PartialView("GetUser", response.Data);
+            }
+
+            var errorModel = ModelState.Values.SelectMany(key => key.Errors.Select(key => key.ErrorMessage));
+            return StatusCode(StatusCodes.Status500InternalServerError, new { errorModel });
+        }
+
         [HttpPost]
         public async Task<IActionResult> DeleteUser(long id)
         {
-            var response = await _userService.DeleteAsync(id);
-            return Ok(response);
+            if (ModelState.IsValid)
+            {
+                var response = await _userService.DeleteAsync(id);
+                if (response.StatusCode == Domain.Enum.StatusCode.Ok)
+                    return RedirectToAction("GetUsers");
+            }
+
+            return RedirectToAction("Index", "Home");
+        }
+
+        [HttpGet]
+        public IActionResult Save()
+        {
+            return PartialView();
         }
 
         [HttpPost]
         public async Task<IActionResult> Save(UserViewModel userViewModel)
         {
-            var user = Enum.Parse<Role>(userViewModel.Role);
             if (ModelState.IsValid)
-                return RedirectToAction("User", "GetUser");
+            {
+                var response = await _userService.CreateUserAsync(userViewModel);
+                if (response.StatusCode == Domain.Enum.StatusCode.Ok)
+                    return Json(new { description = response.Description });
 
-            return Ok(user);
+                return BadRequest(new { errorMessage = response.Description });
+            }
+
+            var errorMessage = ModelState.Values.SelectMany(key => key.Errors.Select(key => key.ErrorMessage)).ToList().Join();
+            return StatusCode(StatusCodes.Status500InternalServerError, new { errorMessage });
+        }
+
+        [HttpPost]
+        public JsonResult GetRoles()
+        {
+            var types = _userService.GetRoles();
+            return Json(types.Data);
         }
     }
 }
